requirements.txt
Flask==2.0.0
numpy==1.22.4
tensorflow==2.12.0
pillow
scikit-learn==1.1.1
nltk=3.7